Combat & Directions
Players can now attack enemies, enemies can attack the player, and both can effectively be "killed".

When either the player or enemies' hp is <= 0, they will be deleted.

Players "direction" changes based on the last movement input, and they can hold down the spacebar to change their direction on the spot without moving. This will not take up a turn.

Players can now input "basic attacks" using the spacebar. This will cast a ray in the direction the player is facing and if it collides with an enemy hitbox, the enemy will be flagged to take damage.